Remove: Support for Elasticsearch Python client & Kerberos authentication
After licence change, Elasticsearch is not being used internally in CERN. Opensearch is used now so only
enable support for this search engine. Likewise, Elasticsearch operations in this code used the Elasticsearch 6.8 cluster
as a reference. Most of them are now deprecated.

Related to Kerberos, AlmaLinux 9 and Lxplus 9 doesn't support gssapi by default. Therefore, this is going to raise issues
with the Requests component for HTTP operations